refactor: replace interfaces
