--- src/app/pages/sprint/interfaces.ts
@@ -1,16 +1,0 @@
-export interface IWordsData {
-  id: string,
-  group: number,
-  page: number,
-  word: string,
-  image: string,
-  audio: string,
-  audioMeaning: string,
-  audioExample: string,
-  textMeaning: string,
-  textExample: string,
-  transcription: string,
-  textExampleTranslate: string,
-  textMeaningTranslate: string,
-  wordTranslate: string
-}
